Split actions from main
